Split casadi backend out into multiple files.
