+++ pymola/backends/sympy/generator.py
@@ -30,7 +30,7 @@ class SympyGenerator(TreeListener):
 from __future__ import print_function, division
 import sympy
 import sympy.physics.mechanics as mech
-from pymola.sympy_runtime import OdeModel
+from pymola.backends.sympy.runtime import OdeModel
 from sympy import sin, cos, tan
 {%- for class_key, class in tree.classes.items() %}
 {{ render.src[class] }}
